requirements_colab.txt
# YOLOX Hazmat Detection - Google Colab Requirements
# Carefully version-locked to prevent dependency conflicts

# CRITICAL: NumPy must be 1.x to avoid PyTorch compatibility issues
numpy==1.24.3

# Computer vision and ML packages compatible with NumPy 1.x
opencv-python==4.8.0.76

# Core ML training dependencies
loguru==0.7.0
tqdm==4.65.0
tabulate==0.9.0
ninja==1.11.1
thop==0.1.1.post2209072238

# COCO evaluation (install with --no-build-isolation to avoid conflicts)
pycocotools>=2.0.6

# ONNX packages - skip initially to avoid conflicts, install separately if needed
# onnx==1.14.0
# onnxruntime==1.15.0
# onnx-simplifier==0.4.36

# PyTorch will be installed separately with CUDA support via conda/pip
# torch>=1.13.1
# torchvision>=0.14.1

# TensorBoard for training visualization (compatible versions)
tensorboard>=2.12.0,<2.15.0

# Additional utilities
Pillow>=9.0.0,<10.0.0
scipy>=1.9.0,<1.12.0
matplotlib>=3.5.0,<4.0.0